feat: 优化门店管理组件，引入naive-ui
