+++ components/stores/Search.vue
@@ -1,6 +1,4 @@
 <script setup lang="ts">
-import type { Rules } from 'common-form'
-
 const props = defineProps<{
   showName: { province_name: string }
 }>()
@@ -10,46 +8,85 @@ const emits = defineEmits<{
   cleanParentId: []
   cleanProvince: []
   updateParent: [val: string]
-  submit: [val: Where]
+  submit: []
 }>()
-const { formList, searchKey } = storeToRefs(useStores())
+const { formList, searchKey, realSearchKey } = storeToRefs(useStores())
 const form = defineModel<Where>({ default: {
 } })
 form.value = formList.value
-
-const rules = ref<Rules<Where>>({
-  parent_id: [],
-  address: [],
-  name: [],
-  province: [],
-  city: [],
-  contact: [],
-  district: [],
-  wxwork_id: [],
-})
-
-// const searchKey = ref<string>('')
-const onSearch = useDebounceFn(() => {
-  emits('updateParent', searchKey.value)
-}, 1000)
-const clearFn = () => {
-  searchKey.value = ''
-  form.value.parent_id = ''
-}
-const blurClean = () => {
-  if (searchKey.value === '') {
-    form.value.parent_id = ''
+const rules = {}
+// form 表单尺寸
+const size = ref<'small' | 'medium' | 'large'>('large')
+// 转换名称
+const toLabel = (key: string): string => {
+  const labelMap: { [key: string]: string } = {
+    address: '地址',
+    parent_id: '上级id',
+    name: '店名',
+    province: '省',
+    city: '市',
+    district: '区',
+    contact: '联系方式',
+    wxwork_id: '企业微信id',
   }
-  if (form.value.parent_id === '') {
-    searchKey.value = ''
-  }
+
+  return labelMap[key] || ''
 }
+const options = ref<{ label: string, key: string }[]>([])
+
 // 显示上级门店列表
 const pop = ref(false)
 const popList = ref<storesList[]>([])
 const showPopup = (res: any) => {
   popList.value = res
+  options.value = []
+
+  res.forEach((item: storesList) => {
+    options.value.push({
+      label: item.name,
+      key: item.id,
+    })
+  })
   pop.value = true
+}
+
+const onSearch = useDebounceFn(() => {
+  if (searchKey.value)
+    emits('updateParent', searchKey.value)
+  else
+    form.value.parent_id = undefined
+}, 1000)
+// 清空按钮
+const clearFn = () => {
+  searchKey.value = ''
+  form.value.parent_id = ''
+}
+
+// 失去焦点清空 输入框
+const blurClean = () => {
+  const resetSearch = () => {
+    searchKey.value = ''
+    realSearchKey.value = ''
+  }
+  // 如果 searchKey 或 parent_id 为空，重置所有相关状态
+  if (searchKey.value === '' || !form.value.parent_id) {
+    form.value.parent_id = ''
+    resetSearch()
+    return
+  }
+  // 如果 parent_id 存在，设置 searchKey 为 realSearchKey
+  searchKey.value = realSearchKey.value
+  // 延迟关闭弹出层
+  setTimeout(() => {
+    pop.value = false
+  }, 100)
+}
+// 选择上级门店
+const seletParent = (key: string | number, options: any) => {
+  form.value.parent_id = options.key as string
+  searchKey.value = options.label as string
+  realSearchKey.value = options.label as string
+  pop.value = false
 }
 defineExpose({
   showPopup,
@@ -58,185 +95,57 @@ defineExpose({
 
 <template>
   <div>
-    <common-form v-model="form" :rules="rules" @submit="(val) => emits('submit', val)">
-      <template
-        #parent_id>
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            上级门店
-          </div>
-          <van-popover v-model:show="pop" trigger="manual" placement="bottom-start">
-            <div class="px-[12px] w-[260px]">
-              <template v-if="popList.length === 0">
-                <div class="py-[12px] text-[14px] color-[#666]">
-                  暂无相关门店,换个关键字试试吧
-                </div>
-              </template>
-
-              <template v-for="(item, index) in popList" :key="index">
-                <div
-                  class="py-[12px] border-b-[#E6E6E8] border-b-solid border text-[14px] overflow-hidden text-ellipsis text-nowrap"
-                  :style="{ borderBottom: index === popList.length - 1 ? 'none' : '1px solid #E6E6E8', color: item.id === form.parent_id ? '#578AFA' : '#000' }"
-                  @click="() => {
-                    searchKey = item.name
-                    form.parent_id = item.id
-                    pop = false
-                  }">
-                  {{ item.name }}
-                </div>
-              </template>
-            </div>
-            <template #reference>
-              <van-search
-                v-model="searchKey"
-                autocomplete="off"
-                shape="round"
-                :style="{
-                  '--van-search-padding': '0',
-                  '--van-search-background': '#ffffff',
-                  '--van-search-content-background': '#fff',
-                }"
-                background="#F1F5FE"
-                placeholder="请输入搜索关键词"
-                @update:model-value="onSearch"
-                @search="onSearch"
-                @clear="clearFn"
-                @blur="blurClean"
-
-              />
-            </template>
-          </van-popover>
+    <n-form
+      :model="form"
+      :rules="rules"
+      :size="size"
+      label-placement="top"
+    >
+      <n-form-item :span="12" label="上级门店" path="address">
+        <n-dropdown :options="options" :show="pop" placement="bottom-start" @select="seletParent">
+          <n-input
+            v-model:value="searchKey"
+            round placeholder="请输入上级门店" clearable @input="() => onSearch()" @clear="clearFn" @blur="blurClean" />
+        </n-dropdown>
+      </n-form-item>
+      <div class="pb-[16px]">
+        <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
+          省市区
         </div>
-      </template>
-      <template
-        #address="{
-          error,
-          validate,
-        }">
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            地址
-          </div>
-          <div class="bg-[#fff] rounded-full px-[12px] flex items-center">
-            <input
-              v-model="form.address" class="py-[10px] border-0 text-[14px] flex-1 color-[#333] " type="text" placeholder="请输入地址" @change="validate()">
-            <van-icon name="arrow" />
-          </div>
-          <template v-if="error">
-            <div class="error">
-              {{ error }}
+        <div class="bg-[#fff] border-[#E2E2E8] border-solid border rounded-full px-[12px] flex items-center">
+          <template v-if="form.province || form.city || form.district">
+            <div class="text-[14px] color-[#333] py-[9.5px] flex-1" @click="emits('selectCity')">
+              {{ props.showName.province_name }}
             </div>
           </template>
-        </div>
-      </template>
-      <template
-        #name="{
-          error,
-          validate,
-        }">
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            店名
-          </div>
-          <div class="bg-[#fff] rounded-full px-[12px] flex items-center">
-            <input
-              v-model="form.name" class="py-[10px] border-0 text-[14px] flex-1 color-[#333] " type="text" placeholder="请输入店名" @change="validate()">
-            <van-icon name="arrow" />
-          </div>
-          <template v-if="error">
-            <div class="error">
-              {{ error }}
+          <template v-else>
+            <div class="text-[14px] color-[#C9C9C9] py-[9.5px] flex-1" @click="emits('selectCity')">
+              请输入省市区
             </div>
           </template>
+          <van-icon
+            name="close" @click="() => {
+              form.province = ''
+              form.city = ''
+              form.district = ''
+              emits('cleanProvince')
+            }" />
         </div>
+      </div>
+      <template v-for="(value, key, index) in form" :key="index">
+        <template v-if="key !== 'parent_id' && key !== 'province' && key !== 'city' && key !== 'district'">
+          <n-form-item :span="12" :label="toLabel(key)" path="address">
+            <n-input v-model:value="form[key]" round :placeholder="`请输入${toLabel(key)}`" clearable />
+          </n-form-item>
+        </template>
       </template>
-      <template
-        #province="{
-          error,
-        }">
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            省市区
-          </div>
-          <div class="bg-[#fff] rounded-full px-[12px] flex items-center">
-            <template v-if="form.province || form.city || form.district">
-              <div class="text-[14px] color-[#333] py-[10px] flex-1" @click="emits('selectCity')">
-                {{ props.showName.province_name }}
-              </div>
-            </template>
-            <template v-else>
-              <div class="text-[14px] color-[#666] py-[10px] flex-1" @click="emits('selectCity')">
-                请输入省市区
-              </div>
-            </template>
-
-            <van-icon
-              name="close" @click="() => {
-                form.province = ''
-                form.city = ''
-                form.district = ''
-                emits('cleanProvince')
-              }" />
-          </div>
-          <template v-if="error">
-            <div class="error">
-              {{ error }}
-            </div>
-          </template>
+      <div
+        class="text-size-[16px] font-semibold" @click="emits('submit')">
+        <div class="ok">
+          确定
         </div>
-      </template>
-
-      <template
-        #contact="{
-          error,
-          validate,
-        }">
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            联系方式
-          </div>
-          <div class="bg-[#fff] rounded-full px-[12px] flex items-center">
-            <input
-              v-model="form.contact" class="py-[10px] border-0 text-[14px] flex-1 color-[#333] " type="text" placeholder="请输入联系方式" :maxlength="11" @change="validate()">
-            <van-icon name="arrow" />
-          </div>
-          <template v-if="error">
-            <div class="error">
-              {{ error }}
-            </div>
-          </template>
-        </div>
-      </template>
-      <template
-        #wxwork_id="{
-          error,
-          validate,
-        }">
-        <div class="pb-[16px]">
-          <div class="text-[14px] color-[#333] line-height-[20px] pb-[8px]">
-            企业微信id
-          </div>
-          <div class="bg-[#fff] rounded-full px-[12px] flex items-center">
-            <input
-              v-model="form.wxwork_id" class="py-[10px] border-0 text-[14px] flex-1 color-[#333] " type="text" placeholder="请输入企业微信id" @change="validate()">
-            <van-icon name="arrow" />
-          </div>
-          <template v-if="error">
-            <div class="error">
-              {{ error }}
-            </div>
-          </template>
-        </div>
-      </template>
-      <template #actions="{ submit }">
-        <div
-          class="text-size-[16px] font-semibold" @click="submit">
-          <div class="ok">
-            确定
-          </div>
-        </div>
-      </template>
-    </common-form>
+      </div>
+    </n-form>
   </div>
 </template>
 
@@ -249,5 +158,9 @@ defineExpose({
 }
 ::v-deep .van-popover__wrapper {
   width: 100%;
+}
+:deep(.n-input:hover) {
+  box-shadow: none !important;
+  border-color: var(--n-border-color) !important; // 或者设置为你想要的颜色
 }
 </style>
